+++ backend/core/types/src/backend-swagger.ts
@@ -1060,6 +1060,110 @@ export class ListingsService {
   }
 }
 
+export class PaperApplicationsService {
+  /**
+   * List paperApplications
+   */
+  list(options: IRequestOptions = {}): Promise<PaperApplication[]> {
+    return new Promise((resolve, reject) => {
+      let url = basePath + '/paperApplications';
+
+      const configs: IRequestConfig = getConfigs('get', 'application/json', url, options);
+
+      let data = null;
+
+      configs.data = data;
+      axios(configs, resolve, reject);
+    });
+  }
+  /**
+   * Create paperApplication
+   */
+  create(
+    params: {
+      /** requestBody */
+      body?: PaperApplicationCreate;
+    } = {} as any,
+    options: IRequestOptions = {}
+  ): Promise<PaperApplication> {
+    return new Promise((resolve, reject) => {
+      let url = basePath + '/paperApplications';
+
+      const configs: IRequestConfig = getConfigs('post', 'application/json', url, options);
+
+      let data = params.body;
+
+      configs.data = data;
+      axios(configs, resolve, reject);
+    });
+  }
+  /**
+   * Update paperApplication
+   */
+  update(
+    params: {
+      /** requestBody */
+      body?: PaperApplicationUpdate;
+    } = {} as any,
+    options: IRequestOptions = {}
+  ): Promise<PaperApplication> {
+    return new Promise((resolve, reject) => {
+      let url = basePath + '/paperApplications/{paperApplicationId}';
+
+      const configs: IRequestConfig = getConfigs('put', 'application/json', url, options);
+
+      let data = params.body;
+
+      configs.data = data;
+      axios(configs, resolve, reject);
+    });
+  }
+  /**
+   * Get paperApplication by id
+   */
+  retrieve(
+    params: {
+      /**  */
+      paperApplicationId: string;
+    } = {} as any,
+    options: IRequestOptions = {}
+  ): Promise<PaperApplication> {
+    return new Promise((resolve, reject) => {
+      let url = basePath + '/paperApplications/{paperApplicationId}';
+      url = url.replace('{paperApplicationId}', params['paperApplicationId'] + '');
+
+      const configs: IRequestConfig = getConfigs('get', 'application/json', url, options);
+
+      let data = null;
+
+      configs.data = data;
+      axios(configs, resolve, reject);
+    });
+  }
+  /**
+   * Delete paperApplication by id
+   */
+  delete(
+    params: {
+      /**  */
+      paperApplicationId: string;
+    } = {} as any,
+    options: IRequestOptions = {}
+  ): Promise<any> {
+    return new Promise((resolve, reject) => {
+      let url = basePath + '/paperApplications/{paperApplicationId}';
+      url = url.replace('{paperApplicationId}', params['paperApplicationId'] + '');
+
+      const configs: IRequestConfig = getConfigs('delete', 'application/json', url, options);
+
+      let data = null;
+
+      configs.data = data;
+      axios(configs, resolve, reject);
+    });
+  }
+}
+
 export class PreferencesService {
   /**
    * List preferences
@@ -2493,12 +2597,29 @@ export interface Asset {
   label: string;
 }
 
+export interface PaperApplication {
+  /**  */
+  language: Language;
+
+  /**  */
+  file?: CombinedFileTypes;
+
+  /**  */
+  id: string;
+
+  /**  */
+  createdAt: Date;
+
+  /**  */
+  updatedAt: Date;
+}
+
 export interface ApplicationMethod {
   /**  */
   type: ApplicationMethodType;
 
   /**  */
-  file?: CombinedFileTypes;
+  paperApplications?: PaperApplication[];
 
   /**  */
   id: string;
@@ -2517,14 +2638,9 @@ export interface ApplicationMethod {
 
   /**  */
   acceptsPostmarkedApplications?: boolean;
-}
-
-export interface AssetCreate {
-  /**  */
-  fileId: string;
 
   /**  */
-  label: string;
+  phoneNumber?: string;
 }
 
 export interface ApplicationMethodCreate {
@@ -2532,7 +2648,7 @@ export interface ApplicationMethodCreate {
   type: ApplicationMethodType;
 
   /**  */
-  file?: CombinedFileTypes;
+  paperApplications?: Id[];
 
   /**  */
   label?: string;
@@ -2542,23 +2658,9 @@ export interface ApplicationMethodCreate {
 
   /**  */
   acceptsPostmarkedApplications?: boolean;
-}
-
-export interface AssetUpdate {
-  /**  */
-  id?: string;
 
   /**  */
-  createdAt?: Date;
-
-  /**  */
-  updatedAt?: Date;
-
-  /**  */
-  fileId: string;
-
-  /**  */
-  label: string;
+  phoneNumber?: string;
 }
 
 export interface ApplicationMethodUpdate {
@@ -2575,7 +2677,7 @@ export interface ApplicationMethodUpdate {
   updatedAt?: Date;
 
   /**  */
-  file?: CombinedFileTypes;
+  paperApplications?: Id[];
 
   /**  */
   label?: string;
@@ -2585,6 +2687,9 @@ export interface ApplicationMethodUpdate {
 
   /**  */
   acceptsPostmarkedApplications?: boolean;
+
+  /**  */
+  phoneNumber?: string;
 }
 
 export interface BooleanInput {
@@ -3237,6 +3342,14 @@ export interface ApplicationUpdate {
 
   /**  */
   submissionDate?: Date;
+}
+
+export interface AssetCreate {
+  /**  */
+  fileId: string;
+
+  /**  */
+  label: string;
 }
 
 export interface CreatePresignedUploadMetadata {
@@ -4183,7 +4296,7 @@ export interface ListingCreate {
   countyCode: CountyCode;
 
   /**  */
-  applicationMethods: ApplicationMethodCreate[];
+  applicationMethods: Id[];
 
   /**  */
   preferences: PreferenceCreate[];
@@ -4404,6 +4517,23 @@ export interface PreferenceUpdate {
   id: string;
 }
 
+export interface AssetUpdate {
+  /**  */
+  id?: string;
+
+  /**  */
+  createdAt?: Date;
+
+  /**  */
+  updatedAt?: Date;
+
+  /**  */
+  fileId: string;
+
+  /**  */
+  label: string;
+}
+
 export interface UnitUpdate {
   /**  */
   amiChart?: CombinedAmiChartTypes;
@@ -4495,7 +4625,7 @@ export interface ListingUpdate {
   updatedAt?: Date;
 
   /**  */
-  applicationMethods: ApplicationMethodUpdate[];
+  applicationMethods: Id[];
 
   /**  */
   preferences: PreferenceUpdate[];
@@ -4688,6 +4818,31 @@ export interface ListingUpdate {
 
   /**  */
   waitlistOpenSpots?: number;
+}
+
+export interface PaperApplicationCreate {
+  /**  */
+  language: Language;
+
+  /**  */
+  file?: CombinedFileTypes;
+}
+
+export interface PaperApplicationUpdate {
+  /**  */
+  language: Language;
+
+  /**  */
+  id?: string;
+
+  /**  */
+  createdAt?: Date;
+
+  /**  */
+  updatedAt?: Date;
+
+  /**  */
+  file?: CombinedFileTypes;
 }
 
 export interface Property {
@@ -5030,7 +5185,7 @@ export enum ApplicationMethodType {
   'ExternalLink' = 'ExternalLink',
   'PaperPickup' = 'PaperPickup'
 }
-export type CombinedFileTypes = AssetUpdate;
+export type CombinedFileTypes = Id;
 export enum InputType {
   'boolean' = 'boolean',
   'text' = 'text',